backend/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tts-api"
version = "1.0.0"
description = "FastAPI backend for text-to-speech conversion using NanoAI TTS"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "TTS Team", email = "support@example.com"}
]

dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-dotenv==1.0.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-multipart==0.0.6",
    "aiofiles==23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.2",
]

[tool.setuptools]
packages = ["app"]
